Added .in-table class for thumbnails in table cells
Made it so as not to use the table-up-4 syntax in order to better
control the size/spacing of the images themselves. If inside rows/cols
then the size of the image will be based on the width of the cell. If
adding max-height to the thumbnail then we have no control of the
spacing. This removes the thumbnails dependency on the cell’s size.